+++ src/texts/errors.ts
@@ -31,7 +31,6 @@ export default {
   'cannot_update_amount_for_erc721': () => `Cannot update amount for ERC721 token`,
   'limits_disabled_for_erc721_or_erc1155': () => `Cannot fetch limits for ERC1155 or ERC721 tokens`,
   'link_only_for_claim': () => 'This link can only be redeemed',
-  'only_original_sender_can_generate_url': () => 'Only original sender can generate new url'
-
-
+  'only_original_sender_can_generate_url': () => 'Only original sender can generate new url',
+  'api_validation_error': () => 'Check the data sent to the server. The server returned an error related to data validation'
 }